+++ CHANGES.md
@@ -45,6 +45,7 @@ Release Notes.
 * Save Envoy http access logs when error occurs.
 * Fix wrong `service_instance_sla` setting in the `topology-instance.yml`.
 * Fix wrong metrics name setting in the `self-observability.yml`.
+* Add telemetry data about metrics in, metrics scraping and trace in metrics to zipkin receiver.
 * Fix tags store of log and trace on h2/mysql/pg storage.
 
 #### UI